+++ modules/beta/emrstudio-on-eks/rbac_stack.py
@@ -255,12 +255,10 @@ class EmrEksRbacStack(Stack):
                 {
                     "id": "AwsSolutions-IAM4",
                     "reason": "Managed Policies are for service account roles only",
-                    "applies_to": "*",
                 },
                 {
                     "id": "AwsSolutions-IAM5",
                     "reason": "Resource access restriced to ADDF resources",
-                    "applies_to": "*",
                 },
             ],
         )