+++ modules/beta/emrstudio-on-eks/studio_stack.py
@@ -433,12 +433,10 @@ class StudioLiveStack(Stack):
                 {
                     "id": "AwsSolutions-IAM4",
                     "reason": "Managed Policies are for service account roles only",
-                    "applies_to": "*",
                 },
                 {
                     "id": "AwsSolutions-IAM5",
                     "reason": "Resource access restriced to ADDF resources",
-                    "applies_to": "*",
                 },
             ],
         )