+++ modules/core/opensearch/stack.py
@@ -21,7 +21,7 @@ import aws_cdk.aws_iam as iam
 import aws_cdk.aws_opensearchservice as opensearch
 import cdk_nag
 from aws_cdk import Aspects, Stack, Tags
-from cdk_nag import NagSuppressions
+from cdk_nag import NagPackSuppression, NagSuppressions
 from constructs import Construct, IConstruct
 
 _logger: logging.Logger = logging.getLogger(__name__)
@@ -119,45 +119,53 @@ class OpenSearchStack(Stack):  # type: ignore
         NagSuppressions.add_stack_suppressions(
             self,
             [
-                {
-                    "id": "AwsSolutions-OS2",
-                    "reason": "Node to Node encryption not enabled - no customer data",
-                    "applies_to": "*",
-                },
-                {
-                    "id": "AwsSolutions-OS3",
-                    "reason": "Access restricted by security group ingress permissions in VPC",
-                    "applies_to": "*",
-                },
-                {
-                    "id": "AwsSolutions-OS4",
-                    "reason": "Single noe for demo purposes",
-                    "applies_to": "*",
-                },
-                {
-                    "id": "AwsSolutions-OS5",
-                    "reason": "Access restricted by security group ingress permissions in VPC",
-                    "applies_to": "*",
-                },
-                {
-                    "id": "AwsSolutions-OS7",
-                    "reason": "Single Node for Demo purposes",
-                    "applies_to": "*",
-                },
-                {
-                    "id": "AwsSolutions-OS8",
-                    "reason": "No customer data - for Demo purposes",
-                    "applies_to": "*",
-                },
-                {
-                    "id": "AwsSolutions-OS9",
-                    "reason": "No logs - for Demo purposes",
-                    "applies_to": "*",
-                },
-                {
-                    "id": "AwsSolutions-IAM4",
-                    "reason": "Managed policies used by service accout roles and managed service",
-                    "applies_to": "*",
-                },
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-OS2",
+                        "reason": "Node to Node encryption not enabled - no customer data",
+                    }
+                ),
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-OS3",
+                        "reason": "Access restricted by security group ingress permissions in VPC",
+                    }
+                ),
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-OS4",
+                        "reason": "Single noe for demo purposes",
+                    }
+                ),
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-OS5",
+                        "reason": "Access restricted by security group ingress permissions in VPC",
+                    }
+                ),
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-OS7",
+                        "reason": "Single Node for Demo purposes",
+                    }
+                ),
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-OS8",
+                        "reason": "No customer data - for Demo purposes",
+                    }
+                ),
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-OS9",
+                        "reason": "No logs - for Demo purposes",
+                    }
+                ),
+                NagPackSuppression(
+                    **{
+                        "id": "AwsSolutions-IAM4",
+                        "reason": "Managed policies used by service accout roles and managed service",
+                    }
+                ),
             ],
         )