+++ modules/demo-only/jupyter-hub/stack.py
@@ -158,12 +158,10 @@ class JupyterHubStack(Stack):  # type: ignore
                 {
                     "id": "AwsSolutions-IAM4",
                     "reason": "Managed Policies are for service account roles only",
-                    "applies_to": "*",
                 },
                 {
                     "id": "AwsSolutions-IAM5",
                     "reason": "Resource access restriced to ADDF resources",
-                    "applies_to": "*",
                 },
             ],
         )