+++ CHANGELOG.md
@@ -5,14 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [V0.2.0] - [UNRELEASED]
+=======
+## [V1.1.0] - [UNRELEASED]
 
 ### **Added**
+
+- Added a pattern for Event bridge triggering StepFunctions triggering AWS Batch
+
+### **Changed**
+- Updated default accountId resolution in sample manifests to simple key:value mapping to reduce confusion
+### **Removed**
+
+=======
+## [V1.0.0] - [09/27/2022]
+
+### **Added**
+
+- updated docs with the information about deploying addf with seedfarmer 2.0v which enables multi region and multi account deployments
 - aws-batch-demo pipeline and manifest
 - image extraction (rosbag to png) pipeline with yolov5 object detection
 - sensor extraction (rosbag to parquet) added to image pipeline
 - added docker module to build container images
 - added lane detection module for building docker image into ECR
+- object and lane detection added to the rosbag-image-pipeline
+- added lane detection WITH YOLOP module for building docker image into ECR
+- replace lanedet with yolop for lane detection pipeline
 
 ### **Changed**
 
@@ -21,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - modifed text in modules/analysis/rosbag-scene-detection
 - enforce HTTPS on OpenSearch
 - EFS FileSystemPolicies to improve security
+- removed all referneces to secrestsmanager in modulestack.yaml where not needed
+- core/eks module added region modifier to masterrole references
 
 ### **Removed**